Move authentication controller to controllers/
