+++ application/__init__.py
@@ -32,7 +32,7 @@ log_handler.set_up_logging(app, GOVUK_ENV)
 
 
 import application.controllers.main
-import application.authentication
+import application.controllers.authentication
 import application.controllers.admin.dashboards
 import admin.registrations
 import application.controllers.admin.transforms
