+++ application/controllers/authentication.py
@@ -1,7 +1,7 @@
 from application import app, csrf
 from flask import redirect, request, session, url_for, flash
 from requests_oauthlib import OAuth2Session
-from helpers import api_permission_required, get_admin_client
+from application.helpers import api_permission_required, get_admin_client
 
 
 def get_authorization_url(session):
